Rename _core to cli
