+++ src/sphinx_autobuild/__main__.py
@@ -1,5 +1,5 @@
 """Entrypoint for `python -m sphinx_autobuild`."""
-from ._core import main
+from .cli import main
 
 if __name__ == "__main__":
     main()
